Move pendulum_driver config to a separate file
